requirements.txt
discord.py==1.0.1
pytz==2019.1
